chore(refactor): cleaned up the starmap components
