+++ app/components/Starmap/StarmapShip.tsx
@@ -1,26 +1,24 @@
-import { useEffect, useMemo, useRef, useState, type RefObject } from "react";
-import { Suspense } from "react";
 import { Line, useGLTF } from "@react-three/drei";
+import { useFrame } from "@react-three/fiber";
+import { clientId, q } from "@thorium/context/AppContext";
+import { useStation } from "@thorium/routes/station/useStation";
+import { useLiveQuery } from "@thorium/utils/live-query/client";
+import { setCursor } from "@thorium/utils/setCursor";
+import type { Meter } from "@thorium/utils/unitTypes";
+import { type RefObject, Suspense, useMemo, useRef } from "react";
 import {
-	CanvasTexture,
 	Color,
 	FrontSide,
 	type Group,
 	type Mesh,
 	type MeshStandardMaterial,
 	Object3D,
-	type Sprite,
 	type Vector3,
 } from "three";
-import { useFrame } from "@react-three/fiber";
-import { useGetStarmapStore } from "./starmapStore";
 import type { Line2 } from "three-stdlib";
-import { clientId, q } from "@thorium/context/AppContext";
-import { useLiveQuery } from "@thorium/utils/live-query/client";
-import { setCursor } from "@thorium/utils/setCursor";
-import type { Meter } from "@thorium/utils/unitTypes";
-import { ConeVisualization } from "@thorium/cards/Targeting/Phasers";
-import { useStation } from "@thorium/routes/station/useStation";
+import { PhasersVisualization } from "./PhasersVisualization";
+import { ShipSprite } from "./ShipSprite";
+import { useGetStarmapStore } from "./starmapStore";
 
 export function StarmapShip({
 	id,
@@ -280,89 +278,4 @@ export function useShipModel(modelAsset: string | undefined) {
 	if (!modelAsset) return null;
 
 	return scene;
-}
-
-const ShipSprite = ({
-	color = "red",
-	spriteAsset,
-	userData,
-	opacity = 1,
-}: {
-	color?: string | number;
-	spriteAsset: string;
-	userData?: any;
-	opacity?: number;
-}) => {
-	const spriteMap = useShipSprite(spriteAsset);
-	const scale = 1 / 50;
-	const ref = useRef<Sprite>(null);
-	useFrame(() => {
-		const isSelected = false;
-		// TODO May 24 2022 - this is used for showing that a ship is selected.
-		// const isSelected = useSelectedShips.getState().selectedIds.includes(id);
-		if (isSelected) {
-			ref.current?.material.color.set("#0088ff");
-		} else {
-			ref.current?.material.color.set(color);
-		}
-	});
-
-	return (
-		<sprite ref={ref} scale={[scale, scale, scale]} userData={userData}>
-			<spriteMaterial
-				attach="material"
-				alphaMap={spriteMap}
-				color={color}
-				opacity={opacity}
-				transparent
-				sizeAttenuation={false}
-				needsUpdate={true}
-				depthTest={true}
-				depthWrite={false}
-			/>
-		</sprite>
-	);
-};
-
-export function useShipSprite(spriteAsset: string) {
-	const canvasDimensions = 2048;
-	const [canvas] = useState(() =>
-		Object.assign(document.createElement("canvas"), {
-			width: canvasDimensions,
-			height: canvasDimensions,
-		}),
-	);
-	const [spriteMap] = useState(() => new CanvasTexture(canvas));
-	useEffect(() => {
-		const ctx = canvas.getContext("2d");
-		const img = new Image();
-		img.src = spriteAsset;
-		img.onload = () => {
-			if (!ctx) return;
-			// Draw the canvas
-			ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
-			const imageData = ctx.getImageData(0, 0, canvas.width, canvas.height);
-			const data = imageData.data;
-			// Convert to black and white
-			for (let i = 0; i < data.length; i += 4) {
-				data[i] = data[i + 1] = data[i + 2] = data[i + 3];
-				// data[i + 3] = 255;
-			}
-			ctx.putImageData(imageData, 0, 0);
-			spriteMap.needsUpdate = true;
-		};
-	}, [spriteAsset, canvas, spriteMap]);
-
-	return spriteMap;
-}
-
-function PhasersVisualization({ shipId }: { shipId: number }) {
-	const [phasers] = q.targeting.phasers.list.useNetRequest({ shipId });
-	return (
-		<>
-			{phasers.map((phaser) => (
-				<ConeVisualization key={phaser.id} {...phaser} />
-			))}
-		</>
-	);
 }
